requirements.txt
aiohttp
dnspython
opensearch-py
psutil
pyrogram
tgcrypto
python-dotenv
telethon
